third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-allowed-expected.txt

ALERT: PASS (1/3)
ALERT: PASS (2/3)
ALERT: PASS (3/3)
